requirements.txt
requests
requests_cache
web3
colorama
tqdm
python-dotenv